+++ src/ui-components/DaysUpdateForm.jsx
@@ -1447,9 +1447,9 @@ export default function DaysUpdateForm(props) {
         step="any"
         value={Sleep}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               Game,
